Refactor: Move Component to Separate File for Improved Modularity and Reusability
